backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
sse-starlette
pandas>=2.0.0
python-multipart>=0.0.6
tabulate>=0.9.0
jupyter-client>=8.0.0
ipykernel>=6.0.0
supabase>=2.0.0
e2b-code-interpreter>=1.0.0
